Moved `Timeline.js` and `Map.js` to frontend/pages
